# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a URL Shortener backend API built with Go.",
        "title": "Shortly API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "tinegagideon@outlook.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://mit-license.org/"
        },
        "version": "0.1.1"
    },
    "host": "https://shortly-5p7d.onrender.com",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Returns pong",
                "summary": "Checks connectivity",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/": {
            "get": {
                "description": "Returns all links",
                "produces": [
                    "application/json"
                ],
                "summary": "Fetch all links",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/shorten": {
            "post": {
                "description": "Returns a Short URL",
                "summary": "Insert an entry for a Short URL and Long URL",
                "parameters": [
                    {
                        "description": "Shorten a Link (custom alias is optional)",
                        "name": "shorten_link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ShortenLink"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Returns a success message",
                "summary": "Delete url data by short url",
                "parameters": [
                    {
                        "description": "Delete a Link",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DeleteLink"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/{link}": {
            "get": {
                "description": "Redirects to the original URL",
                "summary": "Fetch a Original URL by Short URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Redirects to Original URL",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.DeleteLink": {
            "description": "Delete Link Model Url",
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "main.ShortenLink": {
            "description": "Shorten link Model Url, Custom_alias",
            "type": "object",
            "properties": {
                "custom_alias": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}